Initial work to break up large main() function
Currently broken, and I'm not sure of a path forward.

It may be that parsing the arguments and processing the config file are too
interrelated to meaningfully break up, and/or that having a PinboardToolAction
type is overly restrictive, but even so, I should be able to write a much
simpler main() function.